* Fixes for changes in DirectEve v1.330.4904.26383

+++ Questor.Modules/Activities/Traveler.cs
@@ -30,7 +30,7 @@ namespace Questor.Modules.Activities
         //private static DateTime _nextSetEVENavDestination = DateTime.MinValue;
         //private static DateTime _nextGetDestinationPath = DateTime.MinValue;
 
-        private static List<long> _destinationRoute;
+        private static List<int> _destinationRoute;
         public static DirectLocation _location;
         private static IEnumerable<DirectBookmark> myHomeBookmarks;
         private static string _locationName;
